backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
redis==5.0.8
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0
prometheus-client==0.21.0
sentry-sdk==2.18.0
structlog==24.4.0
